+++ src/demo.ts
@@ -1,37 +1,45 @@
-// This is how we create a custom type in TS
-/*
-  This looks like a class in JavaScript. 
-  What is the differences between class and interface ?
-
-  - interfaces strictly exist as a way for you to provide type interface to TS
-  - they are only used at compile time
-  - Interfaces are never available and never appear in your run-time code 
-*/
-
-interface Address {
-  // The question mark meaning the field optional
-  line1?: string;
-  line2?: string;
-  province?: string;
-  region?: string;
-  postalCode: string;
+interface Contact {
+  id: number;
+  name: ContactName;
+  birthDate?: Date;
+  status: ContactStatus;
 }
 
-interface Contact extends Address {
-  id: number;
-  name: string;
-  birthDate?: Date;
+/*
+  what is the problem that enum solve ?
+
+  for example, you have a field named status in an interface
+  the values of status can only be "inactive" or "active"
+
+  the problem is when declaring a new instance of this interface, 
+  developers can mispell the value of status. 
+  
+  and since it is a string, if the developer has "1nactive", the compiler cannot detect the error
+
+  => this is why we have enum
+
+  What is enum ? 
+
+  enum is a special type that has a list of hard-coded list of values
+  enum get compiled at compile time. This allows you to refer enum at runtime
+*/
+
+enum ContactStatus {
+  Active = "active",
+  Inactive = "inactive",
+  New = "new",
 }
 
 let primaryContact: Contact = {
   id: 123,
   name: "Trung",
-  postalCode: "M5A 3K3",
+  status: ContactStatus.Active,
 };
 
-let secondaryContact: Contact = {
-  id: 123,
-  name: "Trung",
-  birthDate: new Date("28-08-1999"),
-  postalCode: "M5A 3K3",
-};
+/*
+  What is type alias ?
+
+  Type alias is simply an alias for another type. They are not a type themself
+  which means they can be used interchangeably with the type that they alias
+*/
+type ContactName = string;